Add platform.GetAWSSession and platform.GetAWSS3Client, refactor GetOrCreateDB into postgresoperator package
